Work on AssemblySwitchParser
