+++ src/main/java/nl/ou/debm/assessor/Assessor.java
@@ -249,7 +249,7 @@ public class Assessor {
         var tempDir = Files.createTempDirectory("debm");
 
         // determine number of threads and setup thread pool
-        int hardwareThreads = Runtime.getRuntime().availableProcessors();
+        int hardwareThreads = 12;//Runtime.getRuntime().availableProcessors();
         if ((iNThreads < 1) || (iNThreads > hardwareThreads)){
             iNThreads = hardwareThreads;
         }
