+++ src/main/java/nl/ou/debm/common/feature5/AssemblySwitchParser.java
@@ -11,16 +11,16 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.regex.Pattern;
 
 public class AssemblySwitchParser {
 
-    private static Pattern _asciiStringPattern = Pattern.compile("\\.asciz\\s+\".+?\"");
+    private static Pattern _asciiStringPattern = Pattern.compile("\\.asciz\\s+\"(.+?)\"");
 
-    public HashMap<Long, Boolean> getSwitchesInfo(IAssessor.CodeInfo ci) {
-        var result = new HashMap<Long, Boolean>();
+    public void setIndirectionInfo(HashMap<Long, SwitchInfo> switchMap, IAssessor.CodeInfo ci) {
         List<String> asmLines = null;
         try {
             asmLines = Files.readAllLines(Paths.get(ci.strAssemblyFilename))
@@ -35,42 +35,52 @@ public class AssemblySwitchParser {
         //These are needed for the assembly helper, but we don't do anything special with these
         var homedRegisters = new ArrayList<String>();
         var registerMap = new HashMap<String, String>();
-        var inSwitchBlock = false;
 
         //Build mapping between string names and values
         var stringValues = new HashMap<String, String>();
         String currentStringLabel = null;
         for (var line : asmLines) {
             var info = ci.compilerConfig.architecture == EArchitecture.X64ARCH ? AssemblyHelper.getX64LineType(line, homedRegisters, registerMap) : AssemblyHelper.getX86LineType(line, registerMap);
-            if (info.type == AsmType.OtherLabel)
-                currentStringLabel = info.value.contains("str.") ? info.value : null;
+            if (info.type == AsmType.StringLabel)
+                currentStringLabel = info.value;
             var asciiStringMatcher = _asciiStringPattern.matcher(line);
-            if (asciiStringMatcher.find()) {
+            if (currentStringLabel != null && asciiStringMatcher.find()) {
                 if (!stringValues.containsKey(currentStringLabel))
                     stringValues.put(currentStringLabel, asciiStringMatcher.group(1));
             }
         }
 
         String valueInRcx = null;
-        Long currentSwitchId = null;
+        SwitchInfo currentSwitch = null;
+        String argumentRegister = ci.compilerConfig.architecture == EArchitecture.X64ARCH ? "%rcx" : "%eax";
+
         for (var line : asmLines) {
             var info = ci.compilerConfig.architecture == EArchitecture.X64ARCH ? AssemblyHelper.getX64LineType(line, homedRegisters, registerMap) : AssemblyHelper.getX86LineType(line, registerMap);
-            if (info.type == AsmType.LoadStringInRegister && info.value2.equals("rcx")) {
-                valueInRcx = info.value;
+
+            //We want to find the code marker before the switch statement.
+            //First, we remember the string in the argument register
+            if (info.type == AsmType.LoadStringInRegister && info.value2.equals(argumentRegister)) {
+                valueInRcx = Arrays.stream(info.value.split("\\(")).findFirst().orElse(null);
                 continue;
             }
-            if (info.type == AsmType.Call && info.value.startsWith("__CM_printf")) {
+
+            //If there is a printf, check if it is a switch start. If so, set current switch
+            if (info.type == AsmType.Call && info.value.toLowerCase().startsWith("__cm_printf")) {
                 assert valueInRcx != null;
                 var printfArgument = stringValues.getOrDefault(valueInRcx, null);
                 if (printfArgument != null && printfArgument.contains(CodeMarker.STRCODEMARKERGUID) && printfArgument.contains(">>" + EFeaturePrefix.INDIRECTIONSFEATURE)) {
                     var marker = new IndirectionsCodeMarker(printfArgument);
-                    currentSwitchId = marker.lngGetSwitchID();
+                    currentSwitch = switchMap.getOrDefault(marker.lngGetSwitchID(), null);
                 }
                 continue;
             }
-            //if jump, check if it is an indirection. After that, clear current switchId
-        }
 
-        return result;
+            //We consider every first jump after a switch code marker as THE switch jump.
+            //Indirection check is simple: if it starts with the dereferencing operator
+            if(currentSwitch != null && info.type == AsmType.Jump){
+                currentSwitch.setImplementedAsIndirection(info.value.startsWith("*"));
+                currentSwitch = null;
+            }
+        }
     }
 }
