+++ src/main/java/nl/ou/debm/common/feature5/IndirectionsLLVMVisitor.java
@@ -3,11 +3,28 @@ package nl.ou.debm.common.feature5;
 import nl.ou.debm.common.antlr.LLVMIRBaseVisitor;
 import nl.ou.debm.common.antlr.LLVMIRParser;
 
+import java.util.HashMap;
+
 public class IndirectionsLLVMVisitor extends LLVMIRBaseVisitor<Object> {
 
+    private final HashMap<Long, SwitchInfo> switches = new HashMap<>(){
+        { put(121L, new SwitchInfo()); }
+    };
     @Override
-    public Object visitSwitchTerm(LLVMIRParser.SwitchTermContext ctx){
+    public Object visitFuncBody(LLVMIRParser.FuncBodyContext ctx){
 
+        for(var block : ctx.basicBlock()){
+
+            for(var instruction : block.instruction()){
+                if(instruction.getText().contains("switch")){
+                    System.out.println("Switch!");
+                }
+            }
+        }
         return null;
+    }
+
+    public HashMap<Long, SwitchInfo> getSwitches() {
+        return switches;
     }
 }
